projeto_agentes_trabalho_final/requirements.txt
fastapi
uvicorn[standard]
spacy
psycopg2-binary
sqlalchemy
sentence-transformers
faiss-cpu
pandas
scikit-learn
torch
